up to Solver and regrupre

+++ include/PlockY/AbstractBlockLoaderFactory.hpp
@@ -16,9 +16,9 @@ namespace PlockY {
     class AbstractBlockFactory {
         static_assert(std::is_arithmetic<Scalar>::value, "Scalar must be a numeric type");
     public:
-        virtual std::unique_ptr<DenseBlock<Scalar>> createDense(const std::string& filePath, int row, int col) = 0;
-        virtual std::unique_ptr<SparseBlock<Scalar>> createSparse(const std::string& filePath, int row, int col) = 0;
-        virtual std::unique_ptr<VecBlock<Scalar>> createVector(const std::string& filePath, int row) = 0;
+        virtual std::unique_ptr<DenseBlock<Scalar>> createDense(const std::string& filePath, size_t row, size_t col) = 0;
+        virtual std::unique_ptr<SparseBlock<Scalar>> createSparse(const std::string& filePath, size_t row, size_t col) = 0;
+        virtual std::unique_ptr<VecBlock<Scalar>> createVector(const std::string& filePath, size_t row) = 0;
 
         virtual ~AbstractBlockFactory() = default;
     };
